Move PDAAAL out of AalWiNes
