+++ src/CMakeLists.txt
@@ -16,11 +16,11 @@ include_directories(${Boost_INCLUDE_DIR})
 
 add_executable(aalwines ${HEADER_FILES} main.cpp model/builders/JuniperBuilder.cpp model/builders/PRexBuilder.cpp
     model/Router.cpp model/RoutingTable.cpp model/Query.cpp model/Network.cpp model/filter.cpp model/NetworkPDAFactory.cpp
-    ${BISON_bparser_OUTPUTS} ${FLEX_flexer_OUTPUTS} query/QueryBuilder.cpp 
-    pdaaal/model/PDA.cpp pdaaal/engine/Moped.cpp pdaaal/engine/PAutomaton.cpp
+    ${BISON_bparser_OUTPUTS} ${FLEX_flexer_OUTPUTS} query/QueryBuilder.cpp
+    pdaaal/engine/Moped.cpp
     utils/parsing.cpp utils/system.cpp)
-add_dependencies(aalwines ptrie-ext rapidxml-ext)
-target_link_libraries(aalwines PRIVATE ${Boost_LIBRARIES})
+add_dependencies(aalwines ptrie-ext rapidxml-ext pdaaal-ext)
+target_link_libraries(aalwines PRIVATE ${Boost_LIBRARIES} pdaaal)
 
 install(TARGETS aalwines
 	RUNTIME DESTINATION bin
