+++ src/model/NetworkPDAFactory.cpp
@@ -25,7 +25,6 @@
  */
 
 #include "NetworkPDAFactory.h"
-#include "pdaaal/model/PDAFactory.h"
 
 namespace aalwines
 {
